requirements.txt
autopep8==1.4.4
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
greenlet==1.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.9.3
pycodestyle==2.5.0
pycparser==2.21
requests==2.22.0
SQLAlchemy==1.4.39
urllib3==1.25.11
Werkzeug==0.16.0
WTForms==2.2.1